refactor: rename custom component helper functions
